gas price print and json friendly typing

+++ RaidsBot/Options/Web3Options.cs
@@ -9,6 +9,6 @@ namespace RaidsBot.Options
         public string RpcUrl { get; set; }
         public string AccountPrivateKey { get; set; }
         public long ChainId { get; set; }
-        public BigInteger GasPrice { get; set; }
+        public long GasPrice { get; set; }
     }
 }
